Change tui_register_info::visible to a method
tui_register_info::visible is redundant with the fact that y==0 means
that the register is not visible.  This patch changes this member in
favor of having a single indication of the register's visibility -- a
method with the same name.  This change makes it clear that
delete_data_content_windows is not needed, so this is removed as well.

Approved-By: Andrew Burgess <aburgess@redhat.com>